+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 * Add CLI tool to eject icons into a specified directory
 * Add Documentation site with examples for every icon
 * Expose SVG sprite of all icons
+* Add icons search filter to the Documentation site